pyproject.toml
[project]
name = "ambient-loop"
version = "1.0.0"
description = "Gmail-Driven Customer Support Agent with LangGraph"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.0",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.0.0",
    "langgraph>=0.4.10",
    "langgraph-checkpoint-postgres",
    "langgraph-cli[inmem]>=0.4",
    "python-dotenv>=1.0.0",
    "openai>=1.5.0",
    "pydantic>=2.4.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "sqlalchemy[asyncio]>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "psycopg[binary]>=3.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    "google-api-python-client==2.88.0",
    "google-auth-httplib2==0.1.0",
    "google-auth-oauthlib==1.2.0",
    "requests==2.31.0",
    "langgraph-sdk",
    "langsmith>=0.0.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent", "db", "erp", "data", "integration", "server", "scripts", "tests"]



